Sebastiaan

+++ server/ftp_controller.py
@@ -7,61 +7,15 @@ from flask import current_app
 
 import base64
 
-def try_to_get_text_file_ftps(file_name, file_type, company_id):
-
-    #session = FTP('145.24.222.235') #Create session with FTP
-    with FTP('145.24.222.235') as session:
-        session.login("Controller", "cC2G'Q_&3qY@=D!@")
-
-        if current_app.config["USING_TEST_FTP"] == True: #If testing == True, we dont want to use regular storage, change to specific ftp storage
-            if "testfolder" not in session.nlst():
-                {"errorCode": 404, "Message": "Test folder does not exist on FTP server"}, 404
-            session.cwd("testfolder")
-
-
-        if f"{company_id}" not in session.nlst(): #Check if company dir exists on FTP Server, if not, return
-            session.quit()
-            return {"errorCode": 404, "Message": "Company directory does not exist on FTP server"}, 404
-
-        session.cwd(f'{company_id}') #Change directory on FTP to the company's
-
-        if file_type not in session.nlst(): #Check if file_type dir exists on FTP Server, if not, return (EG: "templates", "manual")
-            return {"errorCode": 404, "Message": f"Company directory does not contain any {file_type} on FTP server"}, 404
-
-        session.cwd(file_type) #Change directory on FTP to the company's
-        
-        random_file_path = generate_random_path(24, 'html') #Generate random file path for temp storage
-        if path.exists(f'temporary_ftp_storage/{random_file_path}'): #Check for extreme edge case, if path is same as a different parallel request path
-            random_file_path = generate_random_path(24, 'html')
-
-        with open(f'temporary_ftp_storage/{random_file_path}', 'w') as handle: #Create the local (temporary) file
-            if f"{file_name}" not in session.nlst(): #Check if actual file exists on FTP Server in the company id directory, if not, return
-                return {"errorCode": 404, "Message": "Requested file not found on FTP server"}, 404
-            session.retrlines("RETR " + file_name, handle.write) #Retrieve file from FTP server and write it to created temp file
-
-        return_data = io.BytesIO() #Create variable to store the local file contents
-
-        with open(f'temporary_ftp_storage/{random_file_path}', 'rb') as handle: #Write all contents from temporary file to the variable
-            return_data.write(handle.read())
-
-        return_data.seek(0) #return variable position to 0 (after writing it ends up at length-1)
-
-        os.remove(f'temporary_ftp_storage/{random_file_path}') #Remove the temporary_file from local storage    
-        #session.quit() #Close FTP session
-
-        return return_data #Return the extracted Bytes
-
 def try_to_get_file_ftps_binary(file_name, file_type, company_id):
-
     #session = FTP('145.24.222.235') #Create session with FTP
     with FTP('145.24.222.235') as session:
         session.login("Controller", "cC2G'Q_&3qY@=D!@")
 
-        if current_app.config["USING_TEST_FTP"] == True: #If testing == True, we dont want to use regular storage, change to specific ftp storage
+        if current_app.config["USING_TEST_FTP"]: #If testing == True, we dont want to use regular storage, change to specific ftp storage
             if "testfolder" not in session.nlst():
-                {"errorCode": 404, "Message": "Test folder does not exist on FTP server"}, 404
+                return {"errorCode": 404, "Message": "Test folder does not exist on FTP server"}, 404
             session.cwd("testfolder")
-
 
         if f"{company_id}" not in session.nlst(): #Check if company dir exists on FTP Server, if not, return
             return {"errorCode": 404, "Message": "Company directory does not exist on FTP server"}, 404
@@ -91,38 +45,43 @@ def try_to_get_file_ftps_binary(file_name, file_type, company_id):
         return_data.seek(0) #return variable position to 0 (after writing it ends up at length-1)
 
         os.remove(f'temporary_ftp_storage/{random_file_path}') #Remove the temporary_file from local storage    
-        #session.quit() #Close FTP session
 
         return return_data #Return the extracted Bytes
 
-def get_image(file_name, file_type, company_id):
+def try_to_get_image_ftps(file_name, file_type, company_id):
     #session = FTP('145.24.222.235')
     with FTP('145.24.222.235') as session:
         session.login("Controller", "cC2G'Q_&3qY@=D!@")
 
-        if current_app.config["USING_TEST_FTP"] == True: #If testing == True, we dont want to use regular storage, change to specific ftp storage
+        if current_app.config["USING_TEST_FTP"]: #If testing == True, we dont want to use regular storage, change to specific ftp storage
             if "testfolder" not in session.nlst():
-                {"errorCode": 404, "Message": "Test folder does not exist on FTP server"}, 404
+                return {"errorCode": 404, "Message": "Test folder does not exist on FTP server"}, 404
             session.cwd("testfolder")
         
-        session.cwd(f'{company_id}')
-        session.cwd(file_type)
+        if f"{company_id}" not in session.nlst(): #Check if company dir exists on FTP Server, if not, return
+            return {"errorCode": 404, "Message": "Company directory does not exist on FTP server"}, 404
+
+        session.cwd(f'{company_id}') #Change directory on FTP to the company's
+
+        if file_type not in session.nlst(): #Check if file_type dir exists on FTP Server, if not, return (EG: "templates", "manual")
+            return {"errorCode": 404, "Message": f"Company directory does not contain any {file_type} on FTP server"}, 404
+
+        session.cwd(file_type) #Change directory on FTP to the company's
+        
+        random_file_path = generate_random_path(24, 'jpg') #Generate random file path for temp storage
+        if path.exists(f'temporary_ftp_storage/{random_file_path}'): #Check for extreme edge case, if path is same as a different parallel request path
+            random_file_path = generate_random_path(24, 'jpg')
+
+        with open(f'temporary_ftp_storage/{random_file_path}', 'wb') as handle:
+            session.retrbinary("RETR " + file_name, handle.write)
         img = ""
-        try:
-            random_file_path = generate_random_path(24, 'jpg') #Generate random file path for temp storage
-            if path.exists(f'temporary_ftp_storage/{random_file_path}'): #Check for extreme edge case, if path is same as a different parallel request path
-                random_file_path = generate_random_path(24, 'jpg')
-            session.retrbinary("RETR " + file_name, open(f'temporary_ftp_storage/{random_file_path}', 'wb').write)
-            #session.quit()
-            with open(f'temporary_ftp_storage/{random_file_path}', 'rb') as f:
-                for data in f:
-                    img_b64 = base64.b64encode(data)
-                    img_str = img_b64.decode('ascii')
-                    img += img_str
-            os.remove(f'temporary_ftp_storage/{random_file_path}')
-            return img
-        except:
-            return img
+        with open(f'temporary_ftp_storage/{random_file_path}', 'rb') as f:
+            for data in f:
+                img_b64 = base64.b64encode(data)
+                img_str = img_b64.decode('ascii')
+                img = img + img_str
+        os.remove(f'temporary_ftp_storage/{random_file_path}')
+        return img
 
 #Copies a company's template into a company's product directory on FTP
 def try_to_copy_template_to_product(template_name, company_id):
@@ -130,9 +89,9 @@ def try_to_copy_template_to_product(template_name, company_id):
     with FTP('145.24.222.235') as session:
         session.login("Controller", "cC2G'Q_&3qY@=D!@")
 
-        if current_app.config["USING_TEST_FTP"] == True: #If testing == True, we dont want to use regular storage, change to specific ftp storage
+        if current_app.config["USING_TEST_FTP"]: #If testing == True, we dont want to use regular storage, change to specific ftp storage
             if "testfolder" not in session.nlst():
-                {"errorCode": 404, "Message": "Test folder does not exist on FTP server"}, 404
+                return {"errorCode": 404, "Message": "Test folder does not exist on FTP server"}, 404
             session.cwd("testfolder")
 
         if f"{company_id}" not in session.nlst(): #Check if company dir exists on FTP Server, if not, return
@@ -162,28 +121,22 @@ def try_to_copy_template_to_product(template_name, company_id):
         session.cwd("products")
         with open(f'temporary_ftp_storage/{random_file_path}', 'rb') as file_to_send:
             session.storbinary("STOR " + template_name , file_to_send)
-        #session.quit()
+
         os.remove(f"temporary_ftp_storage/{random_file_path}")
         return {"Code": 201, "Message": "Product succesfully created"}, 201
 
 
-#upload_file () name of file, file type (gallery OR templates)
-def upload_file(file_path, file_name, file_type, company_id):
+#try_to_upload_file_ftps () name of file, file type (gallery OR templates)
+def try_to_upload_file_ftps(file_path, file_name, file_type, company_id):
 
     #session = FTP('145.24.222.235') #Create session with FTP
     with FTP('145.24.222.235') as session:
         session.login("Controller", "cC2G'Q_&3qY@=D!@") #Login to FTP
 
-        if current_app.config["USING_TEST_FTP"] == True: #If testing == True, we dont want to use regular storage, change to specific ftp storage
+        if current_app.config["USING_TEST_FTP"]: #If testing == True, we dont want to use regular storage, change to specific ftp storage
             if "testfolder" not in session.nlst():
                 session.mkd("testfolder")
             session.cwd("testfolder")
-
-
-        try:
-            file_to_send = open(file_path, 'rb') #Open file to send
-        except:
-            return {"errorCode": 404, "Message": "The sent file could not be found"}, 404
 
         if f"{company_id}" not in session.nlst(): #Check if company dir exists on FTP Server, if not, create it
             session.mkd(f"{company_id}")
@@ -192,18 +145,19 @@ def upload_file(file_path, file_name, file_type, company_id):
             session.mkd(file_type)
 
         session.cwd(file_type) #Change to the gallery/templates dir
-        session.storbinary("STOR " + file_name, file_to_send) #Send file through as binary
-        #session.quit() #Close FTP session
-        file_to_send.close()
-        return {"Code": 201, "Message": "The file has been uploaded to the FTP server"}, 201
 
-def delete_file_ftps(file_path, file_type, company_id):
+        with open(file_path, 'rb') as file_to_send: #Open file to send
+            session.storbinary("STOR " + file_name, file_to_send) #Send file through as binary
+        #os.remove(file_path)
+        return "PASSED"
+
+def try_to_delete_file_ftps(file_path, file_type, company_id):
     #session = FTP('145.24.222.235')
     with FTP('145.24.222.235') as session:
         session.login("Controller", "cC2G'Q_&3qY@=D!@")
-        if current_app.config["USING_TEST_FTP"] == True: #If testing == True, we dont want to use regular storage, change to specific ftp storage
+        if current_app.config["USING_TEST_FTP"]: #If testing == True, we dont want to use regular storage, change to specific ftp storage
             if "testfolder" not in session.nlst():
-                {"errorCode": 404, "Message": "Test folder does not exist on FTP server"}, 404
+                return {"errorCode": 404, "Message": "Test folder does not exist on FTP server"}, 404
             session.cwd("testfolder")
 
         if f"{company_id}" not in session.nlst(): #Check if company dir exists on FTP Server, if not, return
@@ -220,61 +174,30 @@ def delete_file_ftps(file_path, file_type, company_id):
             return {"errorCode": 404, "Message": f"File {file_path}  does not exist on FTP server"}, 404
 
         session.delete(file_path)
-        #session.quit()
         return "PASSED"
 
-
-def remove_ftp_dir(session, path):
-    for (name, properties) in session.mlsd(path=path):
-        if name in ['.', '..']:
-            continue
-        elif properties['type'] == 'file':
-            session.delete(f"{path}/{name}")
-        elif properties['type'] == 'dir':
-            remove_ftp_dir(session, f"{path}/{name}")
-    session.rmd(path)
-
 def delete_test_folder_ftp():
-    #session = FTP('145.24.222.235')
     with FTP('145.24.222.235') as session:
         session.login("Controller", "cC2G'Q_&3qY@=D!@")
         if "testfolder" not in session.nlst():
             return 1
 
         deletedir("testfolder", session)
-        #session.cwd("testfolder")
-        #for file in session.nlst():
-        #session.delete(file)
-
-        #session.cwd("../")
-        #session.delete("testfolder")
         return 0
 
 def deletedir(dirname, ftp):
-
     ftp.cwd(dirname)
-
-    print(dirname)
-
     for file in ftp.nlst():
-
         try:
             ftp.delete(file)
-
         except Exception:
             deletedir(file, ftp)
-
     ftp.cwd("..")
     ftp.rmd(dirname)
 
-
-
-
-
-
 #try_to_download_text_file(filename, 1)
-#get_image("picture.jpg")
-#upload_file("template1.html", "templates")
+#try_to_get_image_ftps("picture.jpg")
+#try_to_upload_file_ftps("template1.html", "templates")
 #a = getFileContents(filename)
 #ftp.dir()
 #delete_file("picture.jpg")
